refactor: move TritonServerInstanceManager to internal package
