+++ kura/org.eclipse.kura.ai.triton.server/src/main/java/org/eclipse/kura/ai/triton/server/TritonServerServiceRemoteImpl.java
@@ -12,6 +12,7 @@
  ******************************************************************************/
 package org.eclipse.kura.ai.triton.server;
 
+import org.eclipse.kura.ai.triton.server.internal.TritonServerInstanceManager;
 import org.eclipse.kura.executor.CommandExecutorService;
 
 public class TritonServerServiceRemoteImpl extends TritonServerServiceAbs {
